Create code
In this example, we're using WebSockets to handle the real-time communication between users. When a user submits a message through the form, we send it to the WebSocket server using socket.send(). When the WebSocket receives a message from the server, we display it in the chat window using the displayMessage() function.